--- a/lxrhash.go
+++ b/lxrhash.go
@@ -25,6 +25,8 @@ func (lx *LXRHash) UByte(v uint64) uint64 {
 	return uint64(lx.ByteMap[v&lx.MapMask])
 }
 
+// faststep is a simpler step that aims to intialize different start values for different length inputs.
+// prevents someone from pre-calculating a common prefix
 func (lx *LXRHash) faststep(hs []uint64, as, s1, s2, s3, v2, idx uint64) (uint64, uint64, uint64, uint64) {
 	b := lx.UByte(as ^ v2)
 	as = as<<7 ^ as>>5 ^ v2<<20 ^ v2<<16 ^ v2 ^ b<<20 ^ b<<12 ^ b<<4
@@ -33,24 +35,38 @@ func (lx *LXRHash) faststep(hs []uint64, as, s1, s2, s3, v2, idx uint64) (uint64
 	return as, s3, s1, s2
 }
 
+// step is a set of shifts intended to give the hash its pseudorandom attribute.
+// Shifts are not random, they are selected to ensure that prior bytes pulled
+// from the ByteMap contribute to the next access of the ByteMap, either by
+// contributing to the lower bits of the index, or in the upper bits that
+// move the access further in the map.
+//
+// We also pay attention not only to where the ByteMap bits are applied,
+// but what bits we use in the indexing of the ByteMap
+//
+// Tests run against this set of shifts show that the bytes pulled from the
+// ByteMap are evenly distributed over possible byte values (0-255) and indexes
+// into the ByteMap are also evenly distributed, and the deltas between bytes
+// provided map to a curve expected (fewer maximum and minimum deltas, and
+// most deltas around zero.)
 func (lx *LXRHash) step(hs []uint64, as, s1, s2, s3, v2, idx uint64) (uint64, uint64, uint64, uint64) {
-	s1 = s1<<9 ^ s1>>1 ^ as ^ lx.UByte(as>>5^v2)<<3      // Shifts are not random.  They are selected to ensure that
-	s1 = s1<<5 ^ s1>>3 ^ lx.UByte(s1^v2)<<7              // Prior bytes pulled from the ByteMap contribute to the
-	s1 = s1<<7 ^ s1>>7 ^ lx.UByte(as^s1>>7)<<5           // next access of the ByteMap, either by contributing to
-	s1 = s1<<11 ^ s1>>5 ^ lx.UByte(v2^as>>11^s1)<<27     // the lower bits of the index, or in the upper bits that
-	_ = 0                                                // move the access further in the map.
-	hs[idx] = s1 ^ as ^ hs[idx]<<7 ^ hs[idx]>>13         //
-	_ = 0                                                // We also pay attention not only to where the ByteMap bits
-	as = as<<17 ^ as>>5 ^ s1 ^ lx.UByte(as^s1>>27^v2)<<3 // are applied, but what bits we use in the indexing of
-	as = as<<13 ^ as>>3 ^ lx.UByte(as^s1)<<7             // the ByteMap
-	as = as<<15 ^ as>>7 ^ lx.UByte(as>>7^s1)<<11         //
-	as = as<<9 ^ as>>11 ^ lx.UByte(v2^as^s1)<<3          // Tests run against this set of shifts show that the
-	_ = 0                                                // bytes pulled from the ByteMap are evenly distributed
-	s1 = s1<<7 ^ s1>>27 ^ as ^ lx.UByte(as>>3)<<13       // over possible byte values (0-255) and indexes into
-	s1 = s1<<3 ^ s1>>13 ^ lx.UByte(s1^v2)<<11            // the ByteMap are also evenly distributed, and the
-	s1 = s1<<8 ^ s1>>11 ^ lx.UByte(as^s1>>11)<<9         // deltas between bytes provided map to a curve expected
-	s1 = s1<<6 ^ s1>>9 ^ lx.UByte(v2^as^s1)<<3           // (fewer maximum and minimum deltas, and most deltas around
-	_ = 0                                                // zero.
+	s1 = s1<<9 ^ s1>>1 ^ as ^ lx.UByte(as>>5^v2)<<3
+	s1 = s1<<5 ^ s1>>3 ^ lx.UByte(s1^v2)<<7
+	s1 = s1<<7 ^ s1>>7 ^ lx.UByte(as^s1>>7)<<5
+	s1 = s1<<11 ^ s1>>5 ^ lx.UByte(v2^as>>11^s1)<<27
+
+	hs[idx] = s1 ^ as ^ hs[idx]<<7 ^ hs[idx]>>13
+
+	as = as<<17 ^ as>>5 ^ s1 ^ lx.UByte(as^s1>>27^v2)<<3
+	as = as<<13 ^ as>>3 ^ lx.UByte(as^s1)<<7
+	as = as<<15 ^ as>>7 ^ lx.UByte(as>>7^s1)<<11
+	as = as<<9 ^ as>>11 ^ lx.UByte(v2^as^s1)<<3
+
+	s1 = s1<<7 ^ s1>>27 ^ as ^ lx.UByte(as>>3)<<13
+	s1 = s1<<3 ^ s1>>13 ^ lx.UByte(s1^v2)<<11
+	s1 = s1<<8 ^ s1>>11 ^ lx.UByte(as^s1>>11)<<9
+	s1 = s1<<6 ^ s1>>9 ^ lx.UByte(v2^as^s1)<<3
+
 	as = as<<23 ^ as>>3 ^ s1 ^ lx.UByte(as^v2^s1>>3)<<7
 	as = as<<17 ^ as>>7 ^ lx.UByte(as^s1>>3)<<5
 	as = as<<13 ^ as>>5 ^ lx.UByte(as>>5^s1)<<1
@@ -74,41 +90,33 @@ func (lx *LXRHash) step(hs []uint64, as, s1, s2, s3, v2, idx uint64) (uint64, ui
 
 // Hash takes the arbitrary input and returns the resulting hash of length HashSize
 func (lx *LXRHash) Hash(src []byte) []byte {
-	// Keep the byte intermediate results as int64 values until reduced.
+
+	// intermediate results
 	hs := make([]uint64, lx.HashSize)
-	// as accumulates the state as we walk through applying the source data through the lookup map
-	// and combine it with the state we are building up.
-	var as = lx.Seed
-	// We keep a series of states, and roll them along through each byte of source processed.
-	var s1, s2, s3 uint64
-	// Since MapSize is specified in bits, the index mask is the size-1
 
-	// Define a function to move the state by one byte.  This is not intended to be fast
-	// Requires the previous byte read to process the next byte read.  Forces serial evaluation
-	// and removes the possibility of scheduling byte access.
-	//
-	// (Note that use of _ = 0 in lines below are to keep go fmt from messing with comments on the right of the page)
+	// state variables
+	var as, s1, s2, s3, idx uint64
+	as = lx.Seed
 
-	idx := uint64(0)
-	// Fast spin to prevent caching state
-	for _, v2 := range src {
+	// fast steps
+	for _, v := range src {
 		if idx >= lx.HashSize { // Use an if to avoid modulo math
 			idx = 0
 		}
-		as, s1, s2, s3 = lx.faststep(hs, as, s1, s2, s3, uint64(v2), idx)
+		as, s1, s2, s3 = lx.faststep(hs, as, s1, s2, s3, uint64(v), idx)
 		idx++
 	}
 
 	idx = 0
 	// Actual work to compute the hash
-	for i, v2 := range src {
+	for i := range src {
 		if idx >= lx.HashSize { // Use an if to avoid modulo math
 			idx = 0
 		}
 		if i == 0 {
-			lx.FirstIdx = (as>>5 ^ uint64(v2)) & lx.MapMask
+			lx.FirstIdx = (as>>5 ^ hs[i]) & lx.MapMask
 		}
-		as, s1, s2, s3 = lx.step(hs, as, s1, s2, s3, uint64(v2), idx)
+		as, s1, s2, s3 = lx.step(hs, as, s1, s2, s3, uint64(src[i]), idx)
 		idx++
 	}