--- a/lxrhash_test.go
+++ b/lxrhash_test.go
@@ -5,6 +5,7 @@ package lxr
 import (
 	"bytes"
 	"encoding/hex"
+	"fmt"
 	"testing"
 )
 
@@ -15,30 +16,20 @@ func init() {
 	lx.Init(0xfafaececfafaecec, 30, 256, 5)
 	oprhash = lx.Hash([]byte(`Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc dapibus pretium urna, mollis aliquet elit cursus ac. Sed sodales, erat ut volutpat viverra, ante urna pretium est, non congue augue dui sed purus. Mauris vitae mollis metus. Fusce convallis faucibus tempor. Maecenas hendrerit, urna eu lobortis venenatis, neque leo consequat enim, nec placerat tellus eros quis diam. Donec quis vestibulum eros. Maecenas id vulputate justo. Quisque nec feugiat nisi, lacinia pulvinar felis. Pellentesque habitant sed.`))
 }
-
 func BenchmarkHash(b *testing.B) {
-	b.Run("hash", func(b *testing.B) {
-		nonce := []byte{0, 0}
-		for i := 0; i < b.N; i++ {
-			nonce = nonce[:0]
-			for j := i; j > 0; j = j >> 8 {
-				nonce = append(nonce, byte(j))
+	for i := 0; i < 10; i++ {
+		b.Run(fmt.Sprintf("hash run %d", i), func(b *testing.B) {
+			nonce := []byte{0, 0}
+			for i := 0; i < b.N; i++ {
+				nonce = nonce[:0]
+				for j := i; j > 0; j = j >> 8 {
+					nonce = append(nonce, byte(j))
+				}
+				no := append(oprhash, nonce...)
+				lx.Hash(no)
 			}
-			no := append(oprhash, nonce...)
-			lx.Hash(no)
-		}
-	})
-	b.Run("hash again", func(b *testing.B) {
-		nonce := []byte{0, 0}
-		for i := 0; i < b.N; i++ {
-			nonce = nonce[:0]
-			for j := i; j > 0; j = j >> 8 {
-				nonce = append(nonce, byte(j))
-			}
-			no := append(oprhash, nonce...)
-			lx.Hash(no)
-		}
-	})
+		})
+	}
 }
 
 func TestKnownHashes(t *testing.T) {