Since the algorithm itself is unlikely to be changed, I'm proposing to once again factor out the closures in order to speed up the algorithm and ensure all miners are on an even playing field.

Closures are not that great and right now, a call to Hash will instantiate 4 separate closures (faststep, step, b, B). I made them their own functions. This yields a speed increase of approximately 5%:

Before refactor
```
BenchmarkHash/hash_run_0-8                 10000            153208 ns/op
BenchmarkHash/hash_run_1-8                 10000            152108 ns/op
BenchmarkHash/hash_run_2-8                 10000            151708 ns/op
BenchmarkHash/hash_run_3-8                 10000            152108 ns/op
BenchmarkHash/hash_run_4-8                 10000            151108 ns/op
BenchmarkHash/hash_run_5-8                 10000            153408 ns/op
BenchmarkHash/hash_run_6-8                 10000            150908 ns/op
BenchmarkHash/hash_run_7-8                 10000            152208 ns/op
BenchmarkHash/hash_run_8-8                 10000            151908 ns/op
BenchmarkHash/hash_run_9-8                 10000            151808 ns/op

201379 hashes in 30 seconds with 1 threads
398845 hashes in 30 seconds with 2 threads
589711 hashes in 30 seconds with 3 threads
780025 hashes in 30 seconds with 4 threads
```

after
```
BenchmarkHash/hash_run_0-8                 10000            146608 ns/op
BenchmarkHash/hash_run_1-8                 10000            145308 ns/op
BenchmarkHash/hash_run_2-8                 10000            144308 ns/op
BenchmarkHash/hash_run_3-8                 10000            144408 ns/op
BenchmarkHash/hash_run_4-8                 10000            144508 ns/op
BenchmarkHash/hash_run_5-8                 10000            146308 ns/op
BenchmarkHash/hash_run_6-8                 10000            145308 ns/op
BenchmarkHash/hash_run_7-8                 10000            144908 ns/op
BenchmarkHash/hash_run_8-8                 10000            144808 ns/op
BenchmarkHash/hash_run_9-8                 10000            144608 ns/op

209857 hashes in 30 seconds with 1 threads
414493 hashes in 30 seconds with 2 threads
614482 hashes in 30 seconds with 3 threads
812155 hashes in 30 seconds with 4 threads
```