@PaulSnow 2019-09-09
Runnng for 10 million + hashes using the simMiner, the difference of the refactor is about  25958 hps vs 25680 hps with the difference falling over time.  (I can't easily explain why the refactor seems to do better at first and loses its edge over time).  This is roughly a 1% improvement (25958-25680)/25680 = .0108255...

I do believe that the closures make the code much easier to understand.

[refactor.txt](https://github.com/pegnet/LXRHash/files/3592703/refactor.txt)
[original.txt](https://github.com/pegnet/LXRHash/files/3592706/original.txt)