@PaulSnow 2019-09-09
So I got curious about go 1.13 and its impact.  My previous test was with go 1.12.6.  I installed go 1.13

To my surprise, the refactor dropped in performance (at ~10 million hashes) to 24931 hps while with closures increased to 33364 hps.  This makes the closures (33364-24931)/24931 = 33.825% faster

[original.txt](https://github.com/pegnet/LXRHash/files/3592998/original.txt)
[refactor.txt](https://github.com/pegnet/LXRHash/files/3592999/refactor.txt)