@WhoSoup 2019-09-17
Alright, had some more time to investigate this. One thing I noticed is that in the simMiner, the `total++` counter **is not thread-safe** and will drop hash counts. I wrote a new simMiner in the simMiner2 folder just to benchmark this case. It uses an atomic incrementer and also displays time alongside the hash per second with minimal overhead. Using my code I experienced no significant dropoff apart from what is expected due to the growing nonce length.

Using go1.13 on an i7-2600 in windows:

With Closures: ~45,600 hps
Without Closures: ~48,000 hps

Which turns out to be exactly the 5% yield in my original post which was on go1.12.

However using go1.13 on an i3-7100U on ubuntu, the code with closures turned out to be significantly faster:

With Closures: ~29200 hps
Without Closures: ~22500 hps

Which is roughly the ~30% speedup you described in yours.

Turns out the issue isn't that simple and might be heavily CPU/OS dependent, with windows being at fault here.  If you have a setup available where you can dual boot windows/lunix on the same hardware, it'd be interesting to get the results from that. 

I assume the majority of miners are using some flavor of lunix so we should definitely leave the code as is for now. Attached are my full results.

[ubuntu-i3-closures.txt](https://github.com/pegnet/LXRHash/files/3621741/ubuntu-i3-closures.txt)
[ubuntu-i3-no_closures.txt](https://github.com/pegnet/LXRHash/files/3621742/ubuntu-i3-no_closures.txt)
[win-i7-closures.txt](https://github.com/pegnet/LXRHash/files/3621743/win-i7-closures.txt)
[win-i7-no_closures.txt](https://github.com/pegnet/LXRHash/files/3621744/win-i7-no_closures.txt)