@WhoSoup 2019-09-26
From discord:

**[Prestige IT] Swibb** reported benchmarking differences of ~1% on a dual boot windows/linux on a Ryzen chip. **whatsitslike** reported benchmarking difference of 0.5% on a dual boot win10/ubuntu on a Xeon chip.

That leads me to believe the intel cpu architecture has had several generational improvements to handle closures better.