@PaulSnow 2019-09-26
Interestingly, I took an alternative approach and unrolled the Step and FastStep functions instead of removing closures by creating functions.  Should have been faster.

On my Intel® Core™ i7-6770HQ CPU @ 2.60GHz × 8 ... No.  It is slower by about 10% than with closures with go 1.13

Unrolling just the FastStep does gain about 5%

Bottom line, this hash has a very curious profile.